chore: Move component to the legend dialog folder.
