+++ packages/map-template/src/components/LegendDialog/GetLegendSectionsHeight.js
